Add a grid layout to my homepage
Using border boxes I can now dynamically change the size of the page while maintaining a basic structure. I have a <div> on the left and some pictures on the right, there is meant to be a footer at the bottom, currently the border is displaying in the correct place but the text is below.